django-backend/requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==16.0.0
beautifulsoup4==4.4.1
billiard==3.3.0.23
celery==3.1.18
cffi==1.6.0
cryptography==1.4
cssselect==0.9.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.7
django-toolbelt==0.0.1
enum34==1.1.6
gunicorn==19.6.0
idna==2.1
ipaddress==1.0.16
kombu==3.0.35
lxml==3.6.0
parsel==1.0.2
pdfminer==20140328
psycopg2==2.6.1
pudb==2016.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
pycurl==7.43.0
PyDispatcher==2.0.5
Pygments==2.1.3
pyOpenSSL==16.0.0
python-dateutil==2.5.3
pytz==2016.4
queuelib==1.4.2
redis==2.10.3
requests==2.10.0
Scrapy==1.1.0
service-identity==16.0.0
six==1.10.0
SQLAlchemy==1.0.13
static3==0.7.0
Twisted==16.2.0
urwid==1.3.1
w3lib==1.14.2
wheel==0.24.0
zope.interface==4.2.0
